requirements.txt
streamlit
pillow
numpy
pandas
scikit-learn
xgboost
torch
torchvision
pyzbar
opencv-python
joblib